Test.py
I added dark/light theme. Theres a lil bug in it where you need to press the toggle twice for it to work but that's another days issue. And then yeah. Just made it Object Oriented as a class based implementation. Lmk if you guys need clarification on how smth works.